project/web/requirements.txt
django
psycopg2
django-location-field
django-allauth
six
python-dotenv
requests
PyJWT
cryptography
pytz
python-dateutil